[Refactor] Decouple runtime filter port from worker
