+++ be/src/runtime/exec_env.cpp
@@ -160,7 +160,7 @@ void ExecEnv::_refresh_service_contexts() {
     _runtime_services.batch_write_mgr = _batch_write_mgr;
     _runtime_services.stream_load_executor = _stream_load_executor;
     _runtime_services.small_file_mgr = _small_file_mgr;
-    _runtime_services.runtime_filter_worker = _runtime_filter_worker;
+    _runtime_services.runtime_filter_sender = _runtime_filter_worker;
     _runtime_services.runtime_filter_query_lifecycle = _runtime_filter_worker;
     _runtime_services.runtime_filter_cache = _runtime_filter_cache;
     _runtime_services.profile_report_worker = profile_report_worker();
